refactor(app-router): move year page to app router
